+++ src/OpenTelemetry.Instrumentation.Kusto/Implementation/KustoActivitySource.cs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using System.Diagnostics;
-using OpenTelemetry.Trace;
 
 namespace OpenTelemetry.Instrumentation.Kusto.Implementation;
 
@@ -11,5 +10,5 @@ namespace OpenTelemetry.Instrumentation.Kusto.Implementation;
 /// </summary>
 internal static class KustoActivitySource
 {
-    public static readonly ActivitySource ActivitySource = ActivitySourceFactory.Create(typeof(KustoActivitySource), KustoSemanticConventions.SemanticConventionsVersion);
+    public static readonly ActivitySource ActivitySource = Trace.ActivitySourceFactory.Create(typeof(KustoActivitySource), KustoSemanticConventions.SemanticConventionsVersion);
 }