+++ test/OpenTelemetry.Resources.AWS.Tests/AWSEC2DetectorTests.cs
@@ -15,6 +15,22 @@ public class AWSEC2DetectorTests
         Timeout = TimeSpan.FromSeconds(3),
     };
 
+    public static TheoryData<SemanticConventionVersion> SemanticConventionVersions()
+    {
+        var data = new TheoryData<SemanticConventionVersion>();
+
+#if NET
+        foreach (var version in Enum.GetValues<SemanticConventionVersion>())
+#else
+        foreach (var version in Enum.GetValues(typeof(SemanticConventionVersion)).Cast<SemanticConventionVersion>())
+#endif
+        {
+            data.Add(version);
+        }
+
+        return data;
+    }
+
     public static async Task<bool> IsRunningOnEC2()
     {
         try
@@ -35,23 +51,31 @@ public class AWSEC2DetectorTests
             new OpenTelemetry.AWS.AWSSemanticConventions(
                 SemanticConventionVersion.Latest));
 
-        var attributes = awsEC2Detector.Detect().Attributes;
+        var resource = awsEC2Detector.Detect();
+
+        Assert.NotNull(resource);
+
+        var attributes = resource.Attributes;
+
         if (!await IsRunningOnEC2())
         {
-            Assert.Empty(attributes); // will be null as it's not in ec2 environment
+            Assert.Empty(attributes); // Will be empty as it's not in EC2 environment
+            Assert.Null(resource.SchemaUrl);
         }
         else
         {
             Assert.NotEmpty(attributes);
+            Assert.StartsWith("https://opentelemetry.io/schemas/", resource.SchemaUrl);
         }
     }
 
-    [Fact]
-    public void TestExtractResourceAttributes()
+    [Theory]
+    [MemberData(nameof(SemanticConventionVersions))]
+    public void TestExtractResourceAttributes(SemanticConventionVersion semanticConventionVersion)
     {
         var awsEC2Detector = new AWSEC2Detector(
             new OpenTelemetry.AWS.AWSSemanticConventions(
-                SemanticConventionVersion.Latest));
+                semanticConventionVersion));
 
         var sampleEC2IdentityDocumentModel = new SampleAWSEC2IdentityDocumentModel();
         var hostName = "Test host name";